Project Update: Item and Menu programming
This update further elaborates on how to handle items and equipment, as well as manners to equip and unequip items and how they affect Character stats.

The tutorial used for this study implemented menu functions through Event Listener logic, but I intend to adapt
the code for Button usage and implement that instead for the next update.